Refactor API views into modular package
