+++ backend/api/views/__init__.py
@@ -1,0 +1,44 @@
+"""Expose public API views for the application."""
+
+from .activities import ActivityViewSet
+from .auth import CreateUserView
+from .banking import BankAccountViewSet
+from .common import dashboard_summary, get_currency_options
+from .company import CompanyInfoViewSet, CompanySettingsViewSet
+from .customers import CustomerPaymentViewSet, CustomerViewSet
+from .expenses import ExpenseCategoryViewSet, ExpenseViewSet, profit_and_loss_report
+from .products import ProductViewSet
+from .purchases import PurchaseReturnViewSet, PurchaseViewSet
+from .sales import (
+    OfferViewSet,
+    PaymentViewSet,
+    SaleReturnViewSet,
+    SaleViewSet,
+    sales_report,
+)
+from .suppliers import SupplierPaymentViewSet, SupplierViewSet
+
+__all__ = [
+    'ActivityViewSet',
+    'BankAccountViewSet',
+    'CompanyInfoViewSet',
+    'CompanySettingsViewSet',
+    'CreateUserView',
+    'CustomerPaymentViewSet',
+    'CustomerViewSet',
+    'ExpenseCategoryViewSet',
+    'ExpenseViewSet',
+    'OfferViewSet',
+    'PaymentViewSet',
+    'ProductViewSet',
+    'PurchaseReturnViewSet',
+    'PurchaseViewSet',
+    'SaleReturnViewSet',
+    'SaleViewSet',
+    'SupplierPaymentViewSet',
+    'SupplierViewSet',
+    'dashboard_summary',
+    'get_currency_options',
+    'profit_and_loss_report',
+    'sales_report',
+]
