+++ backend/api/views/auth.py
@@ -1,0 +1,13 @@
+"""Authentication and user management views."""
+
+from django.contrib.auth.models import User
+from rest_framework import generics
+
+from ..serializers import UserSerializer
+
+
+class CreateUserView(generics.CreateAPIView):
+    """Allow creation of new user accounts."""
+
+    queryset = User.objects.all()
+    serializer_class = UserSerializer
