+++ backend/api/views/expenses.py
@@ -1,0 +1,109 @@
+"""Expense related API views and reports."""
+
+from datetime import date
+
+from django.db.models import DecimalField, Sum
+from django.db.models.functions import Coalesce
+from rest_framework import viewsets
+from rest_framework.decorators import api_view, permission_classes
+from rest_framework.permissions import IsAuthenticated
+from rest_framework.response import Response
+
+from ..activity_logger import log_activity
+from ..models import Expense, Sale
+from ..serializers import (
+    ExpenseCategorySerializer,
+    ExpenseSerializer,
+)
+
+
+class ExpenseCategoryViewSet(viewsets.ModelViewSet):
+    """CRUD operations for expense categories."""
+
+    serializer_class = ExpenseCategorySerializer
+    permission_classes = [IsAuthenticated]
+
+    def get_queryset(self):
+        return self.request.user.expense_categories.all().order_by('name')
+
+    def perform_create(self, serializer):
+        instance = serializer.save(created_by=self.request.user)
+        log_activity(self.request.user, 'created', instance)
+
+    def perform_update(self, serializer):
+        instance = serializer.save()
+        log_activity(self.request.user, 'updated', instance)
+
+    def perform_destroy(self, instance):
+        log_activity(self.request.user, 'deleted', instance)
+        instance.delete()
+
+
+class ExpenseViewSet(viewsets.ModelViewSet):
+    """CRUD operations for expenses."""
+
+    serializer_class = ExpenseSerializer
+    permission_classes = [IsAuthenticated]
+
+    def get_queryset(self):
+        return self.request.user.expenses.all().order_by('-expense_date')
+
+    def perform_create(self, serializer):
+        instance = serializer.save(created_by=self.request.user)
+        log_activity(self.request.user, 'created', instance)
+
+    def perform_update(self, serializer):
+        instance = serializer.save()
+        log_activity(self.request.user, 'updated', instance)
+
+    def perform_destroy(self, instance):
+        log_activity(self.request.user, 'deleted', instance)
+        instance.delete()
+
+
+@api_view(['GET'])
+@permission_classes([IsAuthenticated])
+def profit_and_loss_report(request):
+    """Provide a profit and loss report for a given date range."""
+
+    user = request.user
+
+    start_date_str = request.query_params.get('start_date', '2000-01-01')
+    end_date_str = request.query_params.get('end_date', date.today().strftime('%Y-%m-%d'))
+
+    sales_in_range = Sale.objects.filter(
+        created_by=user,
+        sale_date__range=[start_date_str, end_date_str],
+    )
+
+    total_revenue = sales_in_range.aggregate(
+        total=Coalesce(Sum('total_amount'), 0, output_field=DecimalField())
+    )['total']
+
+    expenses_in_range = Expense.objects.filter(
+        created_by=user,
+        expense_date__range=[start_date_str, end_date_str],
+    )
+
+    expenses_by_category = (
+        expenses_in_range.values('category__name')
+        .annotate(total=Sum('amount'))
+        .order_by('category__name')
+    )
+
+    total_expenses = expenses_in_range.aggregate(
+        total=Coalesce(Sum('amount'), 0, output_field=DecimalField())
+    )['total']
+
+    net_profit = total_revenue - total_expenses
+
+    report_data = {
+        'start_date': start_date_str,
+        'end_date': end_date_str,
+        'total_revenue': total_revenue,
+        'total_expenses': total_expenses,
+        'net_profit': net_profit,
+        'expenses_breakdown': list(expenses_by_category),
+    }
+
+    return Response(report_data)
